fix: CI RED — typecheck: minhash.py float() cast; lint: B007 kmeans loop var, E741 test var names, B905 zip strict, SIM114 combine branches, F401 unused imports, RUF059 unused vars

+++ src/general_ludd/probabilistic/minhash.py
@@ -239,4 +239,4 @@ class LSH:
                     del self._buckets[bucket]
 
     def similarity_threshold(self) -> float:
-        return (1.0 / self._bands) ** (1.0 / self._rows)
+        return float((1.0 / self._bands) ** (1.0 / self._rows))
